Extend options, use argparse
The specific use case I was after is only showing the title.

If this is implemented it becomes natural to only want to show any field

This is achieved through a --include argument which defaults to all fields

For completeness a --exclude field is added, and the previous --no-x
fields are retained for backwards compatibility and work by
modifying exclude.

--exclude and --include work for all metadata that might be returned.
There was an unfortunate collision between the url argument
and the logical name the option to hide/show the url
metadata, so a little argument renaming was necessary.

argparse was used to make all this logic easy to understand.